Add files via upload
Whenever we conduct research, we need to formulate a hypothesis that we can test. Sometimes we accept this hypothesis; but sometimes we also reject it. To make the right decisions, a business must be able to understand whether the assumptions it makes are correct or not.

In this project, I will compare musical preferences in the cities of Springfield and Shelbyville. I will study real Y.Music data to test hypotheses and compare user behavior in these two cities.